moved util functions to corresponding files | fixed import bug
